requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-groq

pymupdf
pypdf

sentence-transformers

chromadb

python-dotenv
streamlit

easyocr
torch
torchvision

audio-recorder-streamlit
sounddevice

elevenlabs
requests